+++ src/main/java/org/folio/circulation/rules/Drools.java
@@ -64,9 +64,11 @@ public class Drools {
     kieSession.insert(new LoanType(loanTypeId));
     kieSession.insert(new PatronGroup(patronGroupId));
     kieSession.insert(new ItemLocation(locationId));
-    kieSession.insert(new Institution(location.getInstitutionId()));
-    kieSession.insert(new Campus(location.getCampusId()));
-    kieSession.insert(new Library(location.getLibraryId()));
+    if (location != null) {
+      kieSession.insert(new Institution(location.getInstitutionId()));
+      kieSession.insert(new Campus(location.getCampusId()));
+      kieSession.insert(new Library(location.getLibraryId()));
+    }
 
     return kieSession;
   }